+++ R/badge.R
@@ -16,6 +16,26 @@ use_cran_badge <- function(base_path = ".") {
   invisible(TRUE)
 }
 
+#' Create a Depsy badge
+#'
+#' This prints out the markdown which will display a Depsy "badge", showing the
+#' "percentile overall impact" of the project, powered by
+#' \url{http://www.depsy.org}.
+#'
+#' Depsy only indexes projects that are on CRAN.
+#'
+#' @inheritParams use_template
+#' @export
+use_depsy_badge <- function(base_path = ".") {
+  pkg <- project_name(base_path)
+
+  src <- paste0("http://depsy.org/api/package/cran/",pkg,"/badge.svg")
+  href <- paste0("http://depsy.org/package/r/", pkg)
+  use_badge("Depsy", href, src, base_path = base_path)
+
+  invisible(TRUE)
+}
+
 #' Use a README badge
 #'
 #' @param badge_name Badge name. Used in error message and alt text